Add some basic tests using pytest
