--- tests/test_import.py
@@ -1,6 +1,0 @@
-import fair
-version = fair.__version__
-
-# Are we dealing with v1.1-x?
-def test_import():
-    assert version[:3] == '1.1'
